file_parser_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9
aiofiles==24.1.0
pdfplumber==0.11.4
pymupdf==1.24.10
pdf2image==1.17.0
pytesseract==0.3.13
python-docx==1.1.2
python-pptx==1.0.2
pandas==2.2.3
openpyxl==3.1.5
markdownify==0.13.1
python-multipart==0.0.20
tabulate==0.9.0
mammoth==1.11.0